chore: refactor tab control
